# openapi.yaml
openapi: 3.0.0
info:
  title: Mini Task API
  description: RESTful API for Task Management System with Authentication, Authorization, Rate Limiting & Idempotency
  version: 1.0.0
  contact:
    name: API Support
    email: support@minitaskapi.com
  license:
    name: ISC
servers:
  - url: http://localhost:3000/api/v1
    description: Development Server
  - url: https://api.example.com/api/v1
    description: Production Server
...
paths:
...
  /users:
    get:
      tags:
        - Users
      summary: Get all users (Admin only)
      description: Retrieve list of all users - requires admin role
      operationId: getAllUsers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /users/{id}:
    get:
      tags:
        - Users
      summary: Get user by ID (Admin only)
      operationId: getUserById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Users
      summary: Update user (Admin only)
      operationId: updateUser
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
                role:
                  type: string
                  enum: [user, premium, admin]
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '403':
          $ref: '#/components/responses/Forbidden'

    delete:
      tags:
        - Users
      summary: Delete user (Admin only)
      operationId: deleteUser
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User deleted successfully
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /tasks:
    get:
      tags:
        - Tasks
      summary: Get all tasks
      description: |
        Retrieve tasks with filtering:
        - Users see: own tasks + public tasks
        - Admin sees: all tasks
      operationId: getAllTasks
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, in_progress, completed]
        - name: priority
          in: query
          schema:
            type: string
            enum: [low, medium, high]
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: No tasks found

    post:
      tags:
        - Tasks
      summary: Create new task
      description: |
        Create task with idempotency support.
        
        **Rate Limiting:**
        - User: 100 requests/15 min
        - Premium: 500 requests/15 min
        
        **Priority Rules:**
        - HIGH priority: Premium users & Admin only
        - MEDIUM/LOW: All users
        
        **Required Headers:**
        - Authorization: Bearer {token}
        - Idempotency-Key: {uuid-v4}
        - Content-Type: application/json
      operationId: createTask
      security:
        - bearerAuth: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          description: Unique request identifier (UUID v4)
          schema:
            type: string
            format: uuid
            example: 550e8400-e29b-41d4-a716-446655440000
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
                  minLength: 1
                  example: My Task
                description:
                  type: string
                  example: Task description
                status:
                  type: string
                  enum: [pending, in_progress, completed]
                  default: pending
                priority:
                  type: string
                  enum: [low, medium, high]
                  default: medium
                assignedTo:
                  type: integer
                  description: User ID to assign task
                isPublic:
                  type: boolean
                  default: false
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  task:
                    $ref: '#/components/schemas/Task'
          headers:
            RateLimit-Limit:
              schema:
                type: integer
              description: Request limit
            RateLimit-Remaining:
              schema:
                type: integer
              description: Remaining requests
            RateLimit-Reset:
              schema:
                type: integer
              description: Reset timestamp
        '400':
          description: Bad request (missing Idempotency-Key, validation error)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (Premium required for HIGH priority)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        example: RATE_LIMIT_EXCEEDED
                      message:
                        type: string
                        example: Too many requests. Try again in 5 minutes
                      retryAfter:
                        type: integer
                        example: 300
...
  /tasks/{id}:
    delete:
      tags:
        - Tasks
      summary: Delete task
      description: |
        Only task owner or admin can delete.
      operationId: deleteTask
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Task deleted successfully
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from login
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: John Doe
        email:
          type: string
          format: email
          example: john@example.com
        role:
          type: string
          enum: [user, premium, admin]
          example: user
        isPremium:
          type: boolean
          example: false
        subscriptionExpiry:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
    Task:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: My Task
        description:
          type: string
          example: Task description
        status:
          type: string
          enum: [pending, in_progress, completed]
          example: pending
        priority:
          type: string
          enum: [low, medium, high]
          example: medium
        ownerId:
          type: integer
          example: 1
        assignedTo:
          type: integer
          nullable: true
        isPublic:
          type: boolean
          example: false
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: UNAUTHORIZED
            message:
              type: string
              example: Invalid credentials
            details:
              type: object
            timestamp:
              type: string
              format: date-time
            path:
              type: string
  responses:
...
    Unauthorized:
      description: Unauthorized - Missing or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'